refactor: Reorganize type definitions for server-client integration
- Move game and socket related types to shared directory
- Restructure type files for better maintainability
- Separate shared types for future server-client integration
- Add `Shared` suffix to shared type files
