+++ client/src/hooks/useCoordinateScale.ts
@@ -1,5 +1,5 @@
 import { RefObject, useCallback, useEffect, useRef } from 'react';
-import { Point } from '@/types/canvas.types';
+import { Point } from '@/types/drawingShared.types';
 
 /**
  * 캔버스 크기 변경 시 사용하는 hook입니다.
