+++ client/src/types/canvas.types.ts
@@ -42,18 +42,3 @@ export interface CanvasEventHandlers {
   onTouchEnd?: (e: TouchEvent<HTMLCanvasElement>) => void;
   onTouchCancel?: (e: TouchEvent<HTMLCanvasElement>) => void;
 }
-
-export interface Point {
-  x: number;
-  y: number;
-}
-
-export interface StrokeStyle {
-  color: string;
-  width: number;
-}
-
-export interface DrawingData {
-  points: Point[];
-  style: StrokeStyle;
-}
