+++ client/src/types/drawingShared.types.ts
@@ -1,0 +1,14 @@
+export interface Point {
+  x: number;
+  y: number;
+}
+
+export interface StrokeStyle {
+  color: string;
+  width: number;
+}
+
+export interface DrawingData {
+  points: Point[];
+  style: StrokeStyle;
+}
