--- client/src/types/setting.types.ts
@@ -1,16 +1,0 @@
-export type Setting = {
-  label: string;
-  key: SettingKey;
-  options: SettingOption[];
-};
-
-export type SettingKey = 'roundCount' | 'playerCount' | 'timeLimit';
-
-export type SettingValues = {
-  [K in SettingKey]: number;
-};
-
-export type SettingOption = {
-  id: number;
-  value: number;
-};
